+++ dio/lib/src/headers.dart
@@ -2,17 +2,25 @@ import 'package:http_parser/http_parser.dart';
 
 import 'utils.dart';
 
+/// The signature that iterates header fields.
 typedef HeaderForEachCallback = void Function(String name, List<String> values);
 
+/// The headers class for requests and responses.
 class Headers {
-  // Header field name
+  Headers() : _map = caseInsensitiveKeyMap<List<String>>();
+
+  /// Create the [Headers] from a [Map] instance.
+  Headers.fromMap(Map<String, List<String>> map)
+      : _map = caseInsensitiveKeyMap<List<String>>(
+          map.map((k, v) => MapEntry(k.trim().toLowerCase(), v)),
+        );
+
   static const acceptHeader = 'accept';
   static const contentEncodingHeader = 'content-encoding';
   static const contentLengthHeader = 'content-length';
   static const contentTypeHeader = 'content-type';
   static const wwwAuthenticateHeader = 'www-authenticate';
 
-  // Header field value
   static const jsonContentType = 'application/json';
   static const formUrlEncodedContentType = 'application/x-www-form-urlencoded';
   static const textPlainContentType = 'text/plain';
@@ -23,13 +31,6 @@ class Headers {
   final Map<String, List<String>> _map;
 
   Map<String, List<String>> get map => _map;
-
-  Headers() : _map = caseInsensitiveKeyMap<List<String>>();
-
-  Headers.fromMap(Map<String, List<String>> map)
-      : _map = caseInsensitiveKeyMap<List<String>>(
-          map.map((k, v) => MapEntry(k.trim().toLowerCase(), v)),
-        );
 
   /// Returns the list of values for the header named [name]. If there
   /// is no header with the provided name, [:null:] will be returned.
@@ -82,10 +83,12 @@ class Headers {
     _map.remove(name);
   }
 
+  /// Clearing all fields in the headers.
   void clear() {
     _map.clear();
   }
 
+  /// Whether the headers has no fields.
   bool get isEmpty => _map.isEmpty;
 
   /// Enumerates the headers, applying the function [f] to each