+++ dio/test/transformer_test.dart
@@ -22,22 +22,27 @@ void main() {
   // Regression: https://github.com/cfug/dio/issues/1834
   test('null response body only when the response is JSON', () async {
     final transformer = BackgroundTransformer();
-    final r1 = await transformer.transformResponse(
-      RequestOptions(responseType: ResponseType.json),
-      ResponseBody.fromBytes([], 200),
-    );
-    expect(r1, '');
-    final r2 = await transformer.transformResponse(
-      RequestOptions(responseType: ResponseType.bytes),
-      ResponseBody.fromBytes([], 200),
-    );
-    expect(r2, []);
-    final r3 = await transformer.transformResponse(
-      RequestOptions(responseType: ResponseType.plain),
-      ResponseBody.fromBytes([], 200),
-    );
-    expect(r3, '');
-    final r4 = await transformer.transformResponse(
+    for (final responseType in ResponseType.values) {
+      final response = await transformer.transformResponse(
+        RequestOptions(responseType: responseType),
+        ResponseBody.fromBytes([], 200),
+      );
+      switch (responseType) {
+        case ResponseType.json:
+        case ResponseType.plain:
+          expect(response, '');
+          break;
+        case ResponseType.stream:
+          expect(response, isA<ResponseBody>());
+          break;
+        case ResponseType.bytes:
+          expect(response, []);
+          break;
+        default:
+          throw AssertionError('Unknown response type: $responseType');
+      }
+    }
+    final jsonResponse = await transformer.transformResponse(
       RequestOptions(responseType: ResponseType.json),
       ResponseBody.fromBytes(
         [],
@@ -47,6 +52,6 @@ void main() {
         },
       ),
     );
-    expect(r4, null);
+    expect(jsonResponse, null);
   });
 }