+++ plugins/http2_adapter/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+*None.*
+
+## 2.3.1+1
+
+- Add topics to packages.
+
+## 2.3.1
+
 - Fix cached `initFuture` not remove when throw exception.
 
 ## 2.3.0